refactor save to comera roll
